requirements.txt
tensorflow-gpu==2.6.0
opencv-python~=4.5.4.60
matplotlib~=3.5.0
numpy~=1.19.5
scikit-image
scikit-learn
pillow
tqdm
ipywidgets
scipy~=1.7.3
pandas~=1.3.4